move h5-related functions to h5_prep.py
